WIP: Added BoardConfig - Could also rename PrebootConfig to BoardConfig instead
